E3-CryoFold/requirement.txt
absl-py
aiohttp
antlr4-python3-runtime
appdirs
argon2-cffi
arrow
asttokens
async-lru
attrs
Babel
beautifulsoup4
biopython
bleach
certifi
cffi
click
cryptography
cycler
deepspeed
defusedxml
docker-pycreds
einops
fair-esm
filelock
flash-attn
fonttools
fsspec
GitPython
gmpy2
h5py
httpx
huggingface-hub
idna
importlib-metadata
importlib-resources
ipykernel
ipython
ipywidgets
Jinja2
joblib
jsonschema
jupyter_client
jupyter_core
jupyterlab
jupyterlab_widgets
kiwisolver
lightning-utilities
lmdb
MarkupSafe
matplotlib
matplotlib-inline
mistune
ml-collections
mrcfile
msgpack
nbclient
nbconvert
nest-asyncio
networkx
notebook
numpy
omegaconf
overrides
packaging
pandas
parso
Pillow
protobuf
psutil
pytest
PyYAML
regex
requests
scikit-learn
scipy
seaborn
Send2Trash
six
sympy
tensorboardX
terminado
threadpoolctl
tqdm
traitlets
transformers
torch
torchaudio
torchmetrics
torchvision
torch_scatter
torch_geometric
torch_cluster
tornado
typing_extensions
urllib3
widgetsnbextension
yacs
yarl
zipp